# college-info-api/swagger.yaml
openapi: 3.0.0
info:
  title: College Info API
  version: 1.0.0
  description: REST API for government college directory in Kalburgi division
servers:
  - url: https://your-app-name.onrender.com
    description: Production server
  - url: http://localhost:5000
    description: Development server
paths:
  /colleges:
    get:
      summary: Get all colleges or filter by district/program (PUBLIC)
      parameters:
        - in: query
          name: district
          schema:
            type: string
          description: Filter by district name
        - in: query
          name: program
          schema:
            type: string
          description: Filter by program/course name
      responses:
        '200':
          description: List of colleges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/College'
    post:
      summary: Add a new college (ADMIN ONLY)
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/College'
      responses:
        '201':
          description: College created
        '401':
          description: Unauthorized
  /colleges/export:
    get:
      summary: Export all colleges to Excel (ADMIN ONLY)
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
  /colleges/{id}:
    get:
      summary: Get college by ID (ADMIN ONLY)
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: College details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/College'
        '404':
          description: College not found
        '401':
          description: Unauthorized
    put:
      summary: Update college by ID (ADMIN ONLY)
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/College'
      responses:
        '200':
          description: College updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/College'
        '404':
          description: College not found
        '401':
          description: Unauthorized
    delete:
      summary: Delete college by ID (ADMIN ONLY)
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: College deleted
        '404':
          description: College not found
        '401':
          description: Unauthorized
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
    College:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        district:
          type: string
        address:
          type: string
        location:
          type: object
          properties:
            googleMapsLink:
              type: string
            mapEmbedUrl:
              type: string
        programs:
          type: array
          items:
            $ref: '#/components/schemas/Program'
        facilities:
          type: array
          items:
            type: string
        contact:
          type: object
          properties:
            phone:
              type: string
            email:
              type: string
            website:
              type: string
    Program:
      type: object
      properties:
        name:
          type: string
        cutoff:
          type: number
        eligibility:
          type: string
        medium:
          type: string